pyproject.toml
[project]
name = "qa-auto-engineer-python-project-241"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pyyaml>=6.0.2",
    "ruff>=0.12.10",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gendiff"]

[project.scripts]
gendiff = "gendiff.scripts.gendiff:main"